requirements/base.txt
-r dev.txt 

asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
psycopg==3.1.16
psycopg-binary==3.1.16
sqlparse==0.4.4
typing_extensions==4.9.0
django-cors-headers==4.3.1